third_party/WebKit/LayoutTests/bluetooth/requestDevice/radio-not-present.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return navigator.bluetooth.test.simulateCentral({state: 'absent'})
    .then(() => assert_promise_rejects_with_message(
      requestDeviceWithKeyDown({filters: [{services: ['generic_access']}]}),
      new DOMException('Bluetooth adapter not available.', 'NotFoundError'),
      'Bluetooth adapter is not present.'));
}, 'Reject with NotFoundError if there is no BT radio present.');
</script>
